-Corrected few typos -Deleted unnecessary imports -Deleted unused functions fit_error_angle and calculate_residual_3D_Points from calibrate.py -Moved function match_data from finish_calibration.py to calibrate.py
